+++ BlazorDeviceControl/Shared/Component/ActionsFilter.razor.cs
@@ -1,10 +1,18 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
+using Microsoft.AspNetCore.Components;
+
 namespace BlazorDeviceControl.Shared.Component
 {
-    public partial class SectionActions
+    public partial class ActionsFilter
     {
+        #region Public and private fields and properties
+
+        [Parameter] public string ItemTypeName { get; set; }
+
+        #endregion
+
         #region Public and private methods
 
         //